# src/docs/swagger.yaml
openapi: '3.0.2'
info:
  version: 1.0.0
  title: My Platform Api's
servers:
  - url: http://localhost:3000/api
...
paths:
  /userTest-v1/test:
    get:
      summary: test api
      operationId: testapi
      tags:
        - Test
      security:
        - ApiAccessKeyAuth: []
        - ApiRefreshKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/headerLanguage'
        - $ref: '#/components/parameters/headerDeviceType'
        - $ref: '#/components/parameters/headerPlatform'
        - $ref: '#/components/parameters/headerVersion'
      responses:
        200:
          description: here will be api response summary!
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    type: object
                    properties:
                      getData:
                        type: array
                        items:
                          $ref: '#/components/schemas/CreateDataResponse'

        400:
          $ref: '#/components/responses/BadRequestError'

    post:
      summary: create data sample api
      operationId: createData
      tags:
        - Test
      security:
        - ApiAccessKeyAuth: []
        - ApiRefreshKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/headerLanguage'
        - $ref: '#/components/parameters/headerDeviceType'
        - $ref: '#/components/parameters/headerPlatform'
        - $ref: '#/components/parameters/headerVersion'
      requestBody:
        required: true
        content:
          application/json:
            example:
              username: 'neetesh'
              email: 'neetesh.trivedi@yopmail.com'
              password: 'user@123'
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
                  format: password
                  description: Password of registered user
      responses:
        200:
          description: here will be api response summary!
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    type: object
                    properties:
                      CreateDataResponse:
                        $ref: '#/components/schemas/CreateDataResponse'
        400:
          $ref: '#/components/responses/BadRequestError'
...
  /superAdminProfile-v1/add:
    get:
      summary: get all admins and super admins
      operationId: getSuperAdminAndAdmins
      tags:
        - Super Admins
      security:
        - ApiAccessKeyAuth: []
        - ApiRefreshKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/headerLanguage'
        - $ref: '#/components/parameters/headerDeviceType'
        - $ref: '#/components/parameters/headerPlatform'
        - $ref: '#/components/parameters/headerVersion'
      responses:
        200:
          description: here will be api response summary!
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    type: object
                    properties:
                      getData:
                        type: array
                        items:
                          $ref: '#/components/schemas/superAdminData'

        400:
          $ref: '#/components/responses/BadRequestError'

    post:
      summary: create data sample api
      operationId: createData
      tags:
        - Super Admins
      security:
        - ApiAccessKeyAuth: []
        - ApiRefreshKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/headerLanguage'
        - $ref: '#/components/parameters/headerDeviceType'
        - $ref: '#/components/parameters/headerPlatform'
        - $ref: '#/components/parameters/headerVersion'
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: 'neetesh'
              email: 'neetesh.trivedi@yopmail.com'
              password: 'user@123'
              role: 'admin' # Set the role (e.g., 'admin', 'super_admin', 'moderator')
              permissions:
                - module: 'user_manager'
                  actions:
                    ['user_create', 'user_edit', 'user_view', 'user_list']
                - module: 'admin_manager'
                  actions:
                    ['admin_create', 'admin_edit', 'admin_view', 'admin_create']
            schema:
              type: object
              required:
                - name
                - email
                - password
                - role # 'role' is required to define admin role
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                  format: password
                  description: Password of the registered user
                role:
                  type: string
                  enum: ['super_admin', 'admin', 'moderator'] # Define enum values here
                  description: Role of the user (must be one of 'super_admin', 'admin', or 'moderator')
                permissions:
                  type: array
                  items:
                    type: object
                    properties:
                      module:
                        type: string
                        description: Name of the module (e.g., 'userManagement', 'contentModeration')
                      actions:
                        type: array
                        items:
                          type: string
                          enum: ['create', 'read', 'update', 'delete'] # List of valid actions
      responses:
        200:
          description: here will be api response summary!
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    type: object
                    properties:
                      CreateDataResponse:
                        $ref: '#/components/schemas/superAdminData'
        400:
          $ref: '#/components/responses/BadRequestError'
components:
  parameters:
    headerLanguage:
      name: Accept-Language
      in: header
      required: true
      schema:
        type: string
        default: en
        enum:
          - en
          - ar
    headerDeviceType:
      name: X-MyApp-Device-Type
      in: header
      required: true
      schema:
        type: string
        default: Linux
        enum:
          - Linux
          - Window
    headerPlatform:
      name: X-MyApp-Platform
      in: header
      required: true
      schema:
        type: string
        default: web
        enum:
          - web
          - ios
          - android
    headerVersion:
      name: X-MyApp-Version
      in: header
      required: true
      schema:
        type: string
        default: 1.0.0
        pattern: '^[\d]+\.[\d]+\.[\d]+$'
...
  schemas:
    ApiResponse:
      required:
        - success
        - message
      properties:
        success:
          type: string
        message:
          type: string
        data: {}
    CreateDataResponse:
      type: object
      properties:
        username:
          type: string
          example: 'neetesh'
        email:
          type: string
          format: email
          example: 'neetesh.trivedi@yopmail.com'
        password:
          type: string
          description: 'Hashed password'
          example: '$2b$10$imKCvSGoyARBfjwSoGxic.h0p.qLT2xnegeXVDf6yp6DfCQcAfjna'
...
    superAdminData:
      type: object
      properties:
        name:
          type: string
          example: 'neetesh'
        email:
          type: string
          format: email
          example: 'neetesh.trivedi@yopmail.com'
        password:
          type: string
          format: password
          example: '$2b$10$cNrX4Owihc29ncSUeyrkJuS4xwye/gMts9Er8sko/G4xFGGZG8AmS'
        role:
          type: string
          example: 'superadmin'
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        isActive:
          type: boolean
          example: true
        reportsTo:
          type: array
          items:
            type: string
        _id:
          type: string
          example: '67c9f6f1d11621d3ece37c79'
        created:
          type: string
          format: date-time
          example: '2025-03-06T19:26:41.714Z'
        updated:
          type: string
          format: date-time
          example: '2025-03-06T19:26:41.714Z'
        __v:
          type: integer
          example: 0
    Permission:
      type: object
      properties:
        module:
          type: string
          example: 'user_manager'
        actions:
          type: array
          items:
            type: string
            example: 'full_access'
  securitySchemes:
    ApiAccessKeyAuth:
      type: apiKey
      in: header
      name: x-access
      description: Access Token for authorization
    ApiRefreshKeyAuth:
      type: apiKey
      in: header
      name: x-refresh
      description: Refresh Token for renewing access tokens
  responses:
    BadRequestError:
      description: Request fails validation or doesn't meet all conditions for request
      content:
        application/json:
          schema:
            example:
              success: 'fail'
              message: 'Validation failed'
              data: ''
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  data:
                    type: object
                    nullable: true
...